refactor: 그룹, 초대 페이지 컴포넌트 절대 경로 적용
