+++ frontend/src/pages/GroupInitPage/components/GroupCreateForm/GroupCreateForm.tsx
@@ -2,9 +2,12 @@ import { ChangeEvent, FormEvent, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { createGroup } from '@/api/group';
 import { Group } from '@/types/group';
-import GroupCreateFormNameInput from '../GroupCreateFormNameInput/GroupCreateFormNameInput';
-import GroupCreateFormSubmitButton from '../GroupCreateFormSubmitButton/GroupCreateFormSubmitButton';
-import { StyledForm, StyledInputContainer } from './GroupCreateForm.styles';
+import GroupCreateFormNameInput from '@/pages/GroupInitPage/components/GroupCreateFormNameInput/GroupCreateFormNameInput';
+import GroupCreateFormSubmitButton from '@/pages/GroupInitPage/components/GroupCreateFormSubmitButton/GroupCreateFormSubmitButton';
+import {
+  StyledForm,
+  StyledInputContainer
+} from '@/pages/GroupInitPage/components/GroupCreateForm/GroupCreateForm.styles';
 import { AxiosError } from 'axios';
 
 function GroupCreateForm() {
