+++ frontend/src/pages/GroupInitPage/components/GroupCreateFormNameInput/GroupCreateFormNameInput.tsx
@@ -5,7 +5,7 @@ import Input from '@/components/Input/Input';
 import TextField from '@/components/TextField/TextField';
 import { Group } from '@/types/group';
 import createPlusImg from '@/assets/create-plus.svg';
-import { StyledCreateIcon } from './GroupCreateFormNameInput.styles';
+import { StyledCreateIcon } from '@/pages/GroupInitPage/components/GroupCreateFormNameInput/GroupCreateFormNameInput.styles';
 
 type Props = {
   groupName: Group['name'];
