+++ frontend/src/pages/GroupInitPage/components/GroupCreateFormSubmitButton/GroupCreateFormSubmitButton.tsx
@@ -1,4 +1,4 @@
-import { StyledButton } from './GroupCreateFormSubmitButton.styles';
+import { StyledButton } from '@/pages/GroupInitPage/components/GroupCreateFormSubmitButton/GroupCreateFormSubmitButton.styles';
 
 function GroupCreateFormSubmitButton() {
   return <StyledButton type="submit">생성하기</StyledButton>;
