+++ frontend/src/pages/GroupInitPage/components/GroupParticipateFormSubmitButton/GroupParticipateFormSubmitButton.tsx
@@ -1,4 +1,4 @@
-import { StyledButton } from './GroupParticipateFormSubmitButton.styles';
+import { StyledButton } from '@/pages/GroupInitPage/components/GroupParticipateFormSubmitButton/GroupParticipateFormSubmitButton.styles';
 
 function GroupParticipateFormSubmitButton() {
   return <StyledButton type="submit">참가하기</StyledButton>;
