+++ frontend/src/pages/InvitationPage/components/InvitationContainer/InvitationContainer.tsx
@@ -1,11 +1,14 @@
 import { useEffect, useState } from 'react';
-import { StyledLogo, StyledTitle } from './InvitationContainer.styles';
+import {
+  StyledLogo,
+  StyledTitle
+} from '@/pages/InvitationPage/components/InvitationContainer/InvitationContainer.styles';
 import { AxiosError } from 'axios';
 import { useParams, useNavigate } from 'react-router-dom';
 import Box from '@/components/Box/Box';
 import serviceLogoImg from '@/assets/service-logo.svg';
 import FlexContainer from '@/components/FlexContainer/FlexContainer';
-import InvitationButtons from '../InvitationButtons/InvitationButtons';
+import InvitationButtons from '@/pages/InvitationPage/components/InvitationButtons/InvitationButtons';
 import { getIsJoinedGroup } from '@/api/group';
 import { Group } from '@/types/group';
